BUG-634: Rework KeyAccess to work on maps
This fixes an invalid assumptions about how the TCPMD5 API works. It
turns out each FD has a detabase of various keys on a per-address basis,
so we really need to promote the single key into a map of keys.

Also fixes a bundle misconfiguration which prevented it from being
useful in OSGi environment.

Change-Id: Idcbb26383b4e787d7eb8d5d2ce183f5c5a782b2e
Signed-off-by: Robert Varga <rovarga@cisco.com>